feat: add getSettings methods
