+++ packages/types/src/index.ts
@@ -4,3 +4,4 @@ export * from './snapApi';
 export * from './constants';
 export * from './params';
 export * from './results';
+export * from './snapInterfaces';
