+++ packages/types/src/snapApi.ts
@@ -7,6 +7,7 @@ import {
   SaveVCOptions,
 } from './params';
 import type { QueryVCsRequestResult, SaveVCRequestResult } from './results';
+import { SSIAccountConfig, SSISnapConfig } from './snapInterfaces';
 
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
 export interface SSISnapEventApi {}
@@ -28,4 +29,6 @@ export interface SSISnapApi {
   setVCStore(store: AvailableVCStores, value: boolean): Promise<boolean>;
   getAvailableVCStores(): Promise<string[]>;
   deleteVC(id: string, options?: DeleteVCsOptions): Promise<boolean[]>;
+  getAccountSettings(): Promise<SSIAccountConfig>;
+  getSnapSettings(): Promise<SSISnapConfig>;
 }
